requirements.txt
pandas
pytest
streamlit
sentence-transformers
scikit-learn
python-dotenv
google-generativeai
